refactor to move layout to it's own file
